requirements.txt
SQLAlchemy==0.8.4
bigfloat==0.3.0a2
bitstring==3.1.2
blist==1.3.4
cdecimal==2.3
ecdsa==0.10
gmpy2==2.0.3
isodate==0.4.9
python-bitcoin==0.0.9
python-gflags==2.0
recordtype==1.1
requests==2.2.0
simplejson==3.3.2
six==1.5.2
sqlalchemy-bitcoin==0.0.2
